Create RegisterConttroller
