+++ TeamProject01/src/main/java/com/example/TeamProject01/controller/pageController.java
@@ -29,27 +29,6 @@ public class pageController {
         this.service = service;
     }
 
-    @ResponseBody
-    @GetMapping("/register/idcheck")
-    public int idchecked(@RequestParam("id") String id){
-        int data = service.idCheck(id);
-        return data;
-    }
-
-    @ResponseBody
-    @GetMapping("/register/nicknameCheck")
-    public int nicknameCheck(@RequestParam("nickname") String n_name) {
-        int data = service.nicknameCheck(n_name);
-        return data;
-    }
-
-    @ResponseBody
-    @GetMapping("/register/mailCheck")
-    public int emailCheck(@RequestParam("mail") String email) {
-        int data = service.emailCheck(email);
-        return data;
-    }
-
     @PostMapping("/aboutUs")
     public String goAboutUs(){
         return "contact/aboutUs";
@@ -88,8 +67,14 @@ public class pageController {
         return "delivery/delivery-search";
     }
 
+    // 상품등록
+    @PostMapping("/addlist")
+    public String goAddList(){
+        return "product/addProduct";
+    }
+
     // 관리자용 상품조회
-    @GetMapping("product")
+    @GetMapping("/product")
     public String goproduct(Model model) {
         return "product/manageProduct";
     }
@@ -124,30 +109,4 @@ public class pageController {
         return "myPage/myReview";
     }
 
-    @PostMapping("/members/new") // 슬러쉬를 붙이지 않으면 이전의 URL 뒤에 이어서 URL 이 나옴, 그리고 슬러쉬를 붙이면
-    // Localhost: XXXX 뒤에 이어서 나옴
-    public String create(@Valid MemberForm mf) throws ParseException {
-        // 우리는 사용자가 입력한 name 값을 가지고 객체를 만듦
-        Date date = new Date();
-        long timeInMilliSeconds = date.getTime();
-        java.sql.Date date1 = new java.sql.Date(timeInMilliSeconds);
-        Timestamp timestamp = new Timestamp(System.currentTimeMillis());
-
-        Member m = new Member();
-        m.setEmail01(mf.getEmail01());
-        m.setId01(mf.getId01());
-        m.setPwd01(mf.getPwd01());
-        m.setName01(mf.getName01());
-        m.setN_name(mf.getN_name());
-        m.setAddr01(mf.getAddr01());
-        m.setAddr02(mf.getAddr02());
-        m.setAddr03(mf.getAddr03());
-        m.setR_date01(date1);
-        m.setP_time(timestamp.getTime());
-
-        // DB에 넣어야 함.
-
-        service.join(m); // 여기에다가 memberForm 을 그대로 때려 넣는것은 좀 아님.
-        return "redirect:/";
-    }
 }
