+++ TeamProject01/src/main/resources/templates/member/register.html
@@ -172,7 +172,7 @@
             <div class="col-lg-5 col-md-8 align-item-center">
                 <div class="border border">
                     <h3 class="bg-gray p-4">Register Now</h3>
-                    <form action="members/new" method="post" id="registration" onsubmit="return formSubmit()">
+                    <form action="register/new" method="post" id="registration" onsubmit="return formSubmit()">
                         <fieldset class="p-4">
                             <div><input type="email" placeholder="Email*" class="border p-3 w-100 my-2" id="email01" name="email01" onchange="mailDupleTest()">
                             <label for="email01" id="mailDupTestMsg"></label></div>
