+++ TeamProject01/src/main/resources/templates/product/addProduct.html
@@ -9,22 +9,22 @@
     <title>상품 추가</title>
 
     <!-- FAVICON -->
-    <link href="../../static/img/favicon.png" rel="shortcut icon">
+    <link href="img/favicon.png" rel="shortcut icon">
     <!-- PLUGINS CSS STYLE -->
     <!-- <link href="plugins/jquery-ui/jquery-ui.min.css" rel="stylesheet"> -->
     <!-- Bootstrap -->
-    <link rel="stylesheet" href="../../static/plugins/bootstrap/css/bootstrap.min.css">
-    <link rel="stylesheet" href="../../static/plugins/bootstrap/css/bootstrap-slider.css">
+    <link rel="stylesheet" href="plugins/bootstrap/css/bootstrap.min.css">
+    <link rel="stylesheet" href="plugins/bootstrap/css/bootstrap-slider.css">
     <!-- Font Awesome -->
-    <link href="../../static/plugins/font-awesome/css/font-awesome.min.css" rel="stylesheet">
+    <link href="plugins/font-awesome/css/font-awesome.min.css" rel="stylesheet">
     <!-- Owl Carousel -->
-    <link href="../../static/plugins/slick-carousel/slick/slick.css" rel="stylesheet">
-    <link href="../../static/plugins/slick-carousel/slick/slick-theme.css" rel="stylesheet">
+    <link href="plugins/slick-carousel/slick/slick.css" rel="stylesheet">
+    <link href="plugins/slick-carousel/slick/slick-theme.css" rel="stylesheet">
     <!-- Fancy Box -->
-    <link href="../../static/plugins/fancybox/jquery.fancybox.pack.css" rel="stylesheet">
-    <link href="../../static/plugins/jquery-nice-select/css/nice-select.css" rel="stylesheet">
+    <link href="plugins/fancybox/jquery.fancybox.pack.css" rel="stylesheet">
+    <link href="plugins/jquery-nice-select/css/nice-select.css" rel="stylesheet">
     <!-- CUSTOM CSS -->
-    <link href="../../static/css/style.css" rel="stylesheet">
+    <link href="css/style.css" rel="stylesheet">
 
 
     <!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
@@ -218,7 +218,7 @@
                     <div class="mobile d-flex">
                         <a href="">
                             <!-- Icon -->
-                            <img src="../../static/images/footer/phone-icon.png" alt="mobile-icon">
+                            <img src="images/footer/phone-icon.png" alt="mobile-icon">
                         </a>
                         <p>참여자의 GitHub</p>
                         <p>
